+++ packages/cli/src/lib/utils/os.ts
@@ -1,0 +1,2 @@
+export const appendCmdIfWindows = (cmd: TemplateStringsArray) =>
+  `${cmd}${process.platform === 'win32' ? '.cmd' : ''}`;