+++ packages/vue-cli-plugin-typescript/generator/template/scripts/start-server.js
@@ -1,6 +1,7 @@
 const {spawn} = require('child_process');
 const {resolve, join} = require('path');
 const {config} = require('dotenv');
+const {appendCmdIfWindows} = require('./utils');
 config();
 
 function getEnvVariables() {
@@ -22,7 +23,7 @@ function getEnvVariables() {
 
 function startServer() {
   const serverPath = join(process.cwd(), 'server');
-  const child = spawn('npm', ['run', 'start'], {
+  const child = spawn(appendCmdIfWindows`npm`, ['run', 'start'], {
     stdio: 'inherit',
     env: getEnvVariables(),
     cwd: resolve(serverPath),