+++ packages/cli-e2e/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.15](https://github.com/coveo/cli/compare/v1.0.14...v1.0.15) (2021-04-21)
+
+**Note:** Version bump only for package @coveo/cli-e2e
+
+
+
+
+
+## [1.0.14](https://github.com/coveo/cli/compare/v1.0.13...v1.0.14) (2021-04-20)
+
+**Note:** Version bump only for package @coveo/cli-e2e
+
+
+
+
+
+## [1.0.13](https://github.com/coveo/cli/compare/v1.0.12...v1.0.13) (2021-04-20)
+
+**Note:** Version bump only for package @coveo/cli-e2e
+
+
+
+
+
+## [1.0.12](https://github.com/coveo/cli/compare/v1.0.11...v1.0.12) (2021-04-20)
+
+**Note:** Version bump only for package @coveo/cli-e2e
+
+
+
+
+
+## [1.0.11](https://github.com/coveo/cli/compare/v1.0.10...v1.0.11) (2021-04-20)
+
+**Note:** Version bump only for package @coveo/cli-e2e
+
+
+
+
+
+## [1.0.10](https://github.com/coveo/cli/compare/v1.0.9...v1.0.10) (2021-04-20)
+
+**Note:** Version bump only for package @coveo/cli-e2e
+
+
+
+
+
+## [1.0.9](https://github.com/coveo/cli/compare/v1.0.8...v1.0.9) (2021-04-20)
+
+**Note:** Version bump only for package @coveo/cli-e2e
+
+
+
+
+
+## [1.0.8](https://github.com/coveo/cli/compare/v1.0.7...v1.0.8) (2021-04-20)
+
+**Note:** Version bump only for package @coveo/cli-e2e
+
+
+
+
+
+## [1.0.7](https://github.com/coveo/cli/compare/v1.0.1...v1.0.7) (2021-04-19)
+
+**Note:** Version bump only for package @coveo/cli-e2e
+
+
+
+
+
 # [1.0.0](https://github.com/coveo/cli/compare/v0.8.0...v1.0.0) (2021-04-15)
 
 **Note:** Version bump only for package cli-e2e